+++ example/bin/test_package.dart
@@ -28,6 +28,10 @@ final List<DartedCommand> commandsTree = [
           acceptedMultiOptionValues: 'android,ios,web',
           isMultiOption: true,
           optionsSeparator: ','),
+
+      // - **NEW in 0.1.5** Set an argument as 'Main', This will make it required and positional.
+      // e.g `your_package create the/path/required`
+      DartedArgument(name: 'path', abbreviation: 'p', isMainReq: true),
     ],
     // Now, give it the flags allowed for this specific command, You can allow it to be negated too! (--no-flag)
     flags: [